backend/pyproject.toml
[project]
name = "math-report-backend"
version = "0.1.0"
description = "Math Report Backend API with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "gunicorn",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pgvector",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "google-genai",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "black",
]

[tool.ruff]
# Ruff configuration
target-version = "py311"
line-length = 100

# Enable specific rule sets
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

# Ignore specific rules
lint.ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    "alembic/versions",
]

# Allow autofix for specific rules
lint.fixable = ["ALL"]
lint.unfixable = []

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"alembic/env.py" = ["E402"]  # Allow module level import not at top

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | alembic/versions
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --strict-markers"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
